Initialized UserList.java
